Removed Add/Edit/DeletePersonCommandTests.
