+++ data/addressbook.json
@@ -3,42 +3,42 @@
     "name" : "Alex Yeoh",
     "phone" : "87438807",
     "email" : "alexyeoh@example.com",
-	"nric": "S9678932J",
+    "nric" : "S9678932J",
     "address" : "Blk 108 McNair Road, #09-310",
     "tagged" : [ "friends" ]
   }, {
     "name" : "Bernice Yu",
     "phone" : "99272758",
     "email" : "berniceyu@example.com",
-	"nric": "S9678932J",
+    "nric" : "S9678932J",
     "address" : "24 Lim Teck Boo Rd, #05-01",
     "tagged" : [ "colleagues", "friends" ]
   }, {
     "name" : "Charlotte Oliveiro",
     "phone" : "93210283",
     "email" : "charlotte@example.com",
-	"nric": "S9678932J",
+    "nric" : "S9678932J",
     "address" : "256 Serangoon Central Dr, #01-01",
     "tagged" : [ "neighbours" ]
   }, {
     "name" : "David Li",
     "phone" : "91031282",
     "email" : "lidavid@example.com",
-	"nric": "S9678932J",
+    "nric" : "S9678932J",
     "address" : "210 Serangoon Central, #16-43",
     "tagged" : [ "family" ]
   }, {
     "name" : "Irfan Ibrahim",
     "phone" : "92492021",
     "email" : "irfan@example.com",
-	"nric": "S9678932J",
+    "nric" : "S9678932J",
     "address" : "900 South Woodlands Dr, #03-01",
     "tagged" : [ "classmates" ]
   }, {
     "name" : "Roy Balakrishnan",
     "phone" : "92624417",
     "email" : "royb@example.com",
-	"nric": "S9678932J",
+    "nric" : "S9678932J",
     "address" : "26 Boon Lay Way, #11-31",
     "tagged" : [ "colleagues" ]
   } ]
