+++ src/main/java/seedu/address/logic/commands/EditHealthWorkerCommand.java
@@ -153,12 +153,13 @@ public class EditHealthWorkerCommand extends EditCommand implements HealthWorker
             EditHealthWorkerDescriptor e = (EditHealthWorkerDescriptor) other;
 
             return getName().equals(e.getName())
+                    && getNric().equals(e.getNric())
                     && getPhone().equals(e.getPhone())
                     && getEmail().equals(e.getEmail())
                     && getAddress().equals(e.getAddress())
                     && getTags().equals(e.getTags())
                     && getOrganization().equals(e.getOrganization())
-                    && getSkills().equals(getSkills());
+                    && getSkills().equals(e.getSkills());
         }
     }
 }
