+++ src/main/java/seedu/address/logic/commands/EditPersonCommand.java
@@ -219,6 +219,7 @@ public class EditPersonCommand extends EditCommand implements PersonCommand {
             EditPersonDescriptor e = (EditPersonDescriptor) other;
 
             return getName().equals(e.getName())
+                    && getNric().equals(e.getNric())
                     && getPhone().equals(e.getPhone())
                     && getEmail().equals(e.getEmail())
                     && getAddress().equals(e.getAddress())
